Extract batch job command as a new app.
